requirements.txt
discord.py
aiohttp
apscheduler
python-dateutil
pytz